normal saving

+++ app/src/main/java/com/example/paintit/HelperDB.java
@@ -9,6 +9,9 @@ import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class HelperDB extends SQLiteOpenHelper {
 
     // database finals
@@ -120,7 +123,7 @@ public class HelperDB extends SQLiteOpenHelper {
         db.insert(DEVELOPMENT, null, values);
         db.close();
     }
-    public String getDevelopment(int paintingId, int userId) {
+    public String getDevelopment(int paintingId, long userId) {
         String pixels = null;
         SQLiteDatabase db = this.getReadableDatabase();
 
@@ -171,6 +174,24 @@ public class HelperDB extends SQLiteOpenHelper {
 
         return "";
     }
+//    public List<User> getNumDrawings() {
+//        db = this.getReadableDatabase();
+//        List<User> users = new ArrayList<User>();
+//        String order = COLUMN_USER_ID + " ASC ";
+//        Cursor cursor = db.query(USER_TABLE, user_columns, null, null, null, null, order); // Cursor is our result
+//        if (cursor.getCount() > 0) {
+//            while (cursor.moveToNext()) {
+//                String username = cursor.getString(cursor.getColumnIndex(COLUMN_USERNAME));
+//                String password = cursor.getString(cursor.getColumnIndex(COLUMN_USER_PASSWORD));
+//                long id = cursor.getLong(cursor.getColumnIndex(COLUMN_USER_ID));
+//                User user = new User(id, username, password);
+//                users.add(user);
+//            }
+//        }
+//        cursor.close();
+//        db.close();
+//        return users;
+//    }
 
     public String getPaintingColors(int paintingId) {
         String colors = null;
